+++ noxfile.py
@@ -8,7 +8,7 @@ to match CLI options, so the test is more than a simple "smoke test" like it is 
 """
 import nox
 
-@nox.session(python=["3.9", "3.10", "3.11", "3.12"])
+@nox.session(python=["3.9", "3.10", "3.11", "3.12", "3.13"])
 def smoke_test(session):
     session.install(".")
     session.run("logmerger", "-h")