+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Support for `strace` log files; added "strace" custom timestamp format, which also accepts leading process id
   integer on each line.
 
+- Added SECURITY.md and security vulnerability reporting guidance in README.md.
+
 
 ## [0.8.0] - 2023-12-07